==> 4-Benchmark.Rmd <==
```{r}
polls_without_scenarios = polls %>%
  filter(!multiple_scenarios)

shared = p3_matched %>%
  filter(!multiple_scenarios) %>%
  filter(CD_CARGO != 5) %>%
  inner_join(polls_without_scenarios, by = c(
    'NR_IDENTIFICACAO_PESQUISA' = 'NR_IDENTIFICACAO_PESQUISA',
    'estimulada' = 'estimulada',
    'CD_CARGO' = 'CD_CARGO',
    'vv' = 'vv',
    'NUMERO_CANDIDATO' = 'NUMERO_CANDIDATO'
  )) %>%
  mutate(agree = abs(result.x - result.y) <= 0.2)
```
